+++ packages/agent/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.5](https://github.com/findify/findify-js/compare/@findify/agent@1.6.4...@findify/agent@1.6.5) (2022-05-24)
+
+**Note:** Version bump only for package @findify/agent
+
+
+
+
+
 ## [1.6.4](https://github.com/findify/findify-js/compare/@findify/agent@1.6.3...@findify/agent@1.6.4) (2022-03-18)